# Metrics optimization exercise