- AverageAge:
    - Start: `7138029 ns/op`
    - `51053 ns/op`
- AveragePaymentAmount
    - Start: `50601786 ns/op`
    - `579205 ns/op`
- StdDevPaymentAmount
    - Start: `116563001 ns/op`
    - `6500811 ns/op`